Extract hot swapping code to a separate module
